finishedTheProject?

+++ src/components/RestartAllProgressModal.tsx
@@ -1,9 +1,19 @@
 import styled from "@emotion/styled"
 import { Modal } from "./Modal";
+import { useEffect, useState } from "react";
 
 export const RestartAllProgressModal = () => {
+
+    const [allowToClick, setAllowToClick] = useState(false);
+
+    useEffect(() => {
+
+        setTimeout(() => {
+            setAllowToClick(true);
+        }, 3000);
+    }, []);
     return (
-        <Modal>
+        <Modal narrow>
             <RestartAllProgressModalWrapper>
                 <div className="informationDiv">
 
@@ -11,31 +21,36 @@ export const RestartAllProgressModal = () => {
                     <p>Are you sure you want to restart all progress?</p>
                 </div>
                 <div className="btnsDiv">
-                    <button onClick={() => {
-                        localStorage.clear();
-                        window.location.reload();
+                    <button className={"btnYes" + (allowToClick ? " allowed" : "")} onClick={() => {
+                        if (allowToClick) {
+
+                            localStorage.clear();
+                            window.location.reload();
+                        }
                     }}>Yes</button>
                     <button onClick={
                         () => {
-                    window.location.reload();
+                            window.location.reload();
                         }
                     }>Cancel</button>
-            </div>
-        </RestartAllProgressModalWrapper>
+                </div>
+
+
+
+            </RestartAllProgressModalWrapper>
         </Modal >
     )
 }
-              // if (3s)
+// if (3s)
 
 const RestartAllProgressModalWrapper = styled.div`
-
 
 .informationDiv {
     display: flex;
     flex-direction: column;
     justify-content: space-between;
     align-items: center;
-    /* height */
+    height: 300px;
     h1 {
         color: red;
         margin-bottom: 1rem;
@@ -50,4 +65,35 @@ const RestartAllProgressModalWrapper = styled.div`
     flex-direction: row;
     justify-content: space-between;
 }
-`
+
+button {
+    width: 150px;
+    font-weight: 600;
+    border: none;
+    padding: 10px;
+    border-radius: 5px;
+    cursor: pointer;
+    background-color: #00c3ff;
+}
+
+.btnYes {
+   background-color: rgba(0, 195, 255, 0.33);
+   animation: allowToClickTrans 3s forwards linear;
+}
+
+@keyframes allowToClickTrans {
+    0% {
+        background-color: rgba(150, 177, 185, 0.33);
+    }
+    95% {
+        background-color: rgba(52, 183, 223, 0.7);
+    }
+    100% {
+        background-color: #00c3ff;
+    }
+}
+
+button.allowed:hover {
+    background-color: #00b3e6 !important;
+}
+`;
